Replace HAVE_OPENSSL_EXT with zend_hash_str_exists
This replaces the HAVE_OPENSSL_EXT condition with more reliable
zend_hash_str_exists. If the openssl extension is loaded as a shared
module on *nix, this symbol is also defined but openssl shared object
(openssl.so) might not be loaded at the runtime.

See: https://github.com/php/php-src/pull/14333